db_service/requirements.txt
numpy
grpcio
protobuf
requests
sqlalchemy
pymysql